2.6.0 1.支持互动抽奖解析 2.优化投票和互动抽奖的显示 3.内部代码结构优化

+++ src/main/java/sandtechnology/bilibili/response/user/UserProfile.java
@@ -13,7 +13,7 @@ public class UserProfile {
     }
 
     public static class Info {
-        int uid;
+        long uid;
         @SerializedName("uname")
         String userName;
         @SerializedName("face")
@@ -27,7 +27,7 @@ public class UserProfile {
             return userName;
         }
 
-        public int getUid() {
+        public long getUid() {
             return uid;
         }
     }
